Scheduling: Move Schedule into its own fragment
