+++ res/xml/schedule.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright (C) 2020 The Dirty Unicorns Project
+
+     Licensed under the Apache License, Version 2.0 (the "License");
+     you may not use this file except in compliance with the License.
+     You may obtain a copy of the License at
+
+          http://www.apache.org/licenses/LICENSE-2.0
+
+     Unless required by applicable law or agreed to in writing, software
+     distributed under the License is distributed on an "AS IS" BASIS,
+     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+     See the License for the specific language governing permissions and
+     limitations under the License.
+-->
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <DropDownPreference
+        android:key="theme_schedule"
+        android:title="@string/theme_schedule_title"
+        android:entries="@array/theme_schedule_titles"
+        android:entryValues="@array/theme_schedule_values"/>
+
+    <SwitchPreference
+        android:key="theme_schedule_repeat_daily"
+        android:title="@string/theme_schedule_repeat_title"
+        android:summary="@string/theme_schedule_repeat_summary"
+        android:defaultValue="true" />
+
+    <ListPreference
+        android:key="scheduled_start_theme"
+        android:dialogTitle="@string/theme_type_summary"
+        android:entries="@array/theme_scheduled_titles"
+        android:entryValues="@array/theme_scheduled_values"/>
+
+    <ListPreference
+        android:key="scheduled_end_theme"
+        android:dialogTitle="@string/theme_type_summary"
+        android:entries="@array/theme_scheduled_titles"
+        android:entryValues="@array/theme_scheduled_values"/>
+
+</PreferenceScreen>
